requirements.txt
buslane==0.0.5
ipykernel==6.13.0
ipython==8.2.0
jupyter-client==7.2.2
jupyter-core==4.9.2
kaleido==0.2.1
keras==2.8.0
Keras-Preprocessing==1.1.2
matplotlib-inline==0.1.3
numpy==1.22.3
pandas==1.4.2
Pillow==8.3.2
plotly==5.7.0
ply==3.11
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tf-estimator-nightly==2.8.0.dev2021122109
matplotlib==3.5.1
matplotlib-inline==0.1.3
opencv-python==4.5.5.64
efficientnet==1.1.1
tensorflow_addons==0.16.1
tqdm==4.64.0
nbformat==5.4.0